physics: forces are reliable yay
