+++ Cargo.toml
@@ -24,19 +24,17 @@ dev_native = [
     "bevy/sysinfo_plugin",
 ]
 config-files = ["ron", "bevy_common_assets", "serde"]
-inspect = ["bevy-inspector-egui"]
+inspect = ["bevy-inspector-egui", "bevy_panorbit_camera/bevy_egui"]
 
 [dependencies]
 # core dependencies
 bevy = "0.14.2"
 bevy-trait-query = "0.6.0"
-rand = "0.8.5"
 # physics dependencies
 avian3d = { version = "0.1.2", features = ["debug-plugin"] }
-rayon = "1.10.0"
 # ui dependencies
+bevy_panorbit_camera = { version = "0.20.0" }
 bevy-inspector-egui = { version = "0.27.0", features = ["highlight_changes"], optional = true }
-bevy_prototype_debug_lines = { version = "0.12.0", features = ["3d"] }
 iyes_perf_ui = { version = "0.3.0" }
 # file io dependencies
 bevy_common_assets = { version = "0.11.0", features = ["ron"], optional = true }
